Fix: Add .toDouble() casts in conductPollWithCampaigns method
- Fixed clamp() calls on lines 1247 and 1251 to properly cast to double
- Ensures num type from clamp() is properly converted to double parameter

+++ government_simulator/lib/services/game_logic_service.dart
@@ -1244,11 +1244,11 @@ class GameLogicService {
     );
 
     // 調査結果に反映
-    final adjustedSupport = (basePoll.playerSupport + campaignImpact).clamp(0, 100);
+    final adjustedSupport = (basePoll.playerSupport + campaignImpact).clamp(0.0, 100.0).toDouble();
 
     // キャンペーンがある場合は誤差範囲が縮小（意見が固まる）
     final campaignInfluence = (activeCampaigns.length + counterCampaigns.length) * 0.5;
-    final adjustedMargin = (basePoll.marginOfError * (1 - campaignInfluence / 100)).clamp(1, 10);
+    final adjustedMargin = (basePoll.marginOfError * (1 - campaignInfluence / 100)).clamp(1.0, 10.0).toDouble();
 
     return Poll(
       id: _uuid.v4(),
